refactor(convex): extract in-app notification queries to inAppNotifications module
