+++ convex/_generated/api.d.ts
@@ -20,6 +20,7 @@ import type * as analytics from "../analytics.js";
 import type * as appointments from "../appointments.js";
 import type * as auth from "../auth.js";
 import type * as authz from "../authz.js";
+import type * as barbershopCascade from "../barbershopCascade.js";
 import type * as barbershopMemberServices from "../barbershopMemberServices.js";
 import type * as barbershopMembers from "../barbershopMembers.js";
 import type * as barbershopMetadata from "../barbershopMetadata.js";
@@ -32,6 +33,7 @@ import type * as errors from "../errors.js";
 import type * as geospatial from "../geospatial.js";
 import type * as http from "../http.js";
 import type * as identity from "../identity.js";
+import type * as inAppNotifications from "../inAppNotifications.js";
 import type * as index from "../index.js";
 import type * as invitations from "../invitations.js";
 import type * as invitationsSchema from "../invitationsSchema.js";
@@ -70,6 +72,7 @@ declare const fullApi: ApiFromModules<{
   appointments: typeof appointments;
   auth: typeof auth;
   authz: typeof authz;
+  barbershopCascade: typeof barbershopCascade;
   barbershopMemberServices: typeof barbershopMemberServices;
   barbershopMembers: typeof barbershopMembers;
   barbershopMetadata: typeof barbershopMetadata;
@@ -82,6 +85,7 @@ declare const fullApi: ApiFromModules<{
   geospatial: typeof geospatial;
   http: typeof http;
   identity: typeof identity;
+  inAppNotifications: typeof inAppNotifications;
   index: typeof index;
   invitations: typeof invitations;
   invitationsSchema: typeof invitationsSchema;
